+++ rectools/models/nn/transformer_base.py
@@ -12,9 +12,6 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-# pylint: disable=too-many-lines
-# TODO: move to 3 separate modules
-
 import io
 import typing as tp
 from collections.abc import Callable
@@ -25,16 +22,12 @@ from tempfile import NamedTemporaryFile
 import numpy as np
 import torch
 import typing_extensions as tpe
-from implicit.gpu import HAS_CUDA
 from pydantic import BeforeValidator, PlainSerializer
-from pytorch_lightning import LightningModule, Trainer
-from scipy import sparse
-from torch.utils.data import DataLoader
+from pytorch_lightning import Trainer
 
 from rectools import ExternalIds
 from rectools.dataset.dataset import Dataset, DatasetSchema, DatasetSchemaDict, IdMap
 from rectools.models.base import ErrorBehaviour, InternalRecoTriplet, ModelBase, ModelConfig
-from rectools.models.rank import Distance, ImplicitRanker
 from rectools.types import InternalIdsArray
 from rectools.utils.misc import get_class_or_function_full_path, import_object
 
@@ -45,566 +38,15 @@ from .item_net import (
     ItemNetConstructorBase,
     SumOfEmbeddingsConstructor,
 )
+from .transformer_backbone import TransformerTorchBackbone
 from .transformer_data_preparator import TransformerDataPreparatorBase
+from .transformer_lightning import TransformerLightningModule, TransformerLightningModuleBase
 from .transformer_net_blocks import (
     LearnableInversePositionalEncoding,
     PositionalEncodingBase,
     PreLNTransformerLayers,
     TransformerLayersBase,
 )
-
-
-class TransformerTorchBackbone(torch.nn.Module):
-    """
-    Torch model for encoding user sessions based on transformer architecture.
-
-    Parameters
-    ----------
-    n_blocks : int
-        Number of transformer blocks.
-    n_factors : int
-        Latent embeddings size.
-    n_heads : int
-        Number of attention heads.
-    session_max_len : int
-        Maximum length of user sequence.
-    dropout_rate : float
-        Probability of a hidden unit to be zeroed.
-    use_pos_emb : bool, default True
-        If ``True``, learnable positional encoding will be added to session item embeddings.
-    use_causal_attn : bool, default True
-        If ``True``, causal mask is used in multi-head self-attention.
-    transformer_layers_type : type(TransformerLayersBase), default `PreLNTransformerLayers`
-        Type of transformer layers architecture.
-    item_net_type : type(ItemNetBase), default `IdEmbeddingsItemNet`
-        Type of network returning item embeddings.
-    pos_encoding_type : type(PositionalEncodingBase), default `LearnableInversePositionalEncoding`
-        Type of positional encoding.
-    """
-
-    def __init__(
-        self,
-        n_blocks: int,
-        n_factors: int,
-        n_heads: int,
-        session_max_len: int,
-        dropout_rate: float,
-        use_pos_emb: bool = True,
-        use_causal_attn: bool = True,
-        use_key_padding_mask: bool = False,
-        transformer_layers_type: tp.Type[TransformerLayersBase] = PreLNTransformerLayers,
-        item_net_block_types: tp.Sequence[tp.Type[ItemNetBase]] = (IdEmbeddingsItemNet, CatFeaturesItemNet),
-        item_net_constructor_type: tp.Type[ItemNetConstructorBase] = SumOfEmbeddingsConstructor,
-        pos_encoding_type: tp.Type[PositionalEncodingBase] = LearnableInversePositionalEncoding,
-    ) -> None:
-        super().__init__()
-
-        self.item_model: ItemNetConstructorBase
-        self.pos_encoding = pos_encoding_type(use_pos_emb, session_max_len, n_factors)
-        self.emb_dropout = torch.nn.Dropout(dropout_rate)
-        self.transformer_layers = transformer_layers_type(
-            n_blocks=n_blocks,
-            n_factors=n_factors,
-            n_heads=n_heads,
-            dropout_rate=dropout_rate,
-        )
-        self.use_causal_attn = use_causal_attn
-        self.use_key_padding_mask = use_key_padding_mask
-        self.n_factors = n_factors
-        self.dropout_rate = dropout_rate
-        self.n_heads = n_heads
-
-        self.item_net_block_types = item_net_block_types
-        self.item_net_constructor_type = item_net_constructor_type
-
-    def construct_item_net(self, dataset: Dataset) -> None:
-        """
-        Construct network for item embeddings from dataset.
-
-        Parameters
-        ----------
-        dataset : Dataset
-            RecTools dataset with user-item interactions.
-        """
-        self.item_model = self.item_net_constructor_type.from_dataset(
-            dataset, self.n_factors, self.dropout_rate, self.item_net_block_types
-        )
-
-    def construct_item_net_from_dataset_schema(self, dataset_schema: DatasetSchema) -> None:
-        """
-        Construct network for item embeddings from dataset schema.
-
-        Parameters
-        ----------
-        dataset_schema : DatasetSchema
-            RecTools schema with dataset statistics.
-        """
-        self.item_model = self.item_net_constructor_type.from_dataset_schema(
-            dataset_schema, self.n_factors, self.dropout_rate, self.item_net_block_types
-        )
-
-    @staticmethod
-    def _convert_mask_to_float(mask: torch.Tensor, query: torch.Tensor) -> torch.Tensor:
-        return torch.zeros_like(mask, dtype=query.dtype).masked_fill_(mask, float("-inf"))
-
-    def _merge_masks(
-        self, attn_mask: torch.Tensor, key_padding_mask: torch.Tensor, query: torch.Tensor
-    ) -> torch.Tensor:
-        """
-        Merge `attn_mask` and `key_padding_mask` as a new `attn_mask`.
-        Both masks are expanded to shape ``(batch_size * n_heads, session_max_len, session_max_len)``
-        and combined with logical ``or``.
-        Diagonal elements in last two dimensions are set equal to ``0``.
-        This prevents nan values in gradients for pytorch < 2.5.0 when both masks are present in forward pass of
-        `torch.nn.MultiheadAttention` (https://github.com/pytorch/pytorch/issues/41508).
-
-        Parameters
-        ----------
-        attn_mask:  torch.Tensor. [session_max_len, session_max_len]
-            Boolean causal attention mask.
-        key_padding_mask: torch.Tensor. [batch_size, session_max_len]
-            Boolean padding mask.
-        query: torch.Tensor
-            Query tensor used to acquire correct shapes and dtype for new `attn_mask`.
-
-        Returns
-        -------
-        torch.Tensor. [batch_size * n_heads, session_max_len, session_max_len]
-            Merged mask to use as new `attn_mask` with zeroed diagonal elements in last 2 dimensions.
-        """
-        batch_size, seq_len, _ = query.shape
-
-        key_padding_mask_expanded = self._convert_mask_to_float(  # [batch_size, session_max_len]
-            key_padding_mask, query
-        ).view(
-            batch_size, 1, seq_len
-        )  # [batch_size, 1, session_max_len]
-
-        attn_mask_expanded = (
-            self._convert_mask_to_float(attn_mask, query)  # [session_max_len, session_max_len]
-            .view(1, seq_len, seq_len)
-            .expand(batch_size, -1, -1)
-        )  # [batch_size, session_max_len, session_max_len]
-
-        merged_mask = attn_mask_expanded + key_padding_mask_expanded
-        res = (
-            merged_mask.view(batch_size, 1, seq_len, seq_len)
-            .expand(-1, self.n_heads, -1, -1)
-            .view(-1, seq_len, seq_len)
-        )  # [batch_size * n_heads, session_max_len, session_max_len]
-        torch.diagonal(res, dim1=1, dim2=2).zero_()
-        return res
-
-    def encode_sessions(self, sessions: torch.Tensor, item_embs: torch.Tensor) -> torch.Tensor:
-        """
-        Pass user history through item embeddings.
-        Add positional encoding.
-        Pass history through transformer blocks.
-
-        Parameters
-        ----------
-        sessions :  torch.Tensor
-            User sessions in the form of sequences of items ids.
-        item_embs : torch.Tensor
-            Item embeddings.
-
-        Returns
-        -------
-        torch.Tensor. [batch_size, session_max_len, n_factors]
-            Encoded session embeddings.
-        """
-        session_max_len = sessions.shape[1]
-        attn_mask = None
-        key_padding_mask = None
-
-        timeline_mask = (sessions != 0).unsqueeze(-1)  # [batch_size, session_max_len, 1]
-
-        seqs = item_embs[sessions]  # [batch_size, session_max_len, n_factors]
-        seqs = self.pos_encoding(seqs)
-        seqs = self.emb_dropout(seqs)
-
-        if self.use_causal_attn:
-            attn_mask = ~torch.tril(
-                torch.ones((session_max_len, session_max_len), dtype=torch.bool, device=sessions.device)
-            )
-        if self.use_key_padding_mask:
-            key_padding_mask = sessions == 0
-            if attn_mask is not None:  # merge masks to prevent nan gradients for torch < 2.5.0
-                attn_mask = self._merge_masks(attn_mask, key_padding_mask, seqs)
-                key_padding_mask = None
-
-        seqs = self.transformer_layers(seqs, timeline_mask, attn_mask, key_padding_mask)
-        return seqs
-
-    def forward(
-        self,
-        sessions: torch.Tensor,  # [batch_size, session_max_len]
-    ) -> tp.Tuple[torch.Tensor, torch.Tensor]:
-        """
-        Forward pass to get item and session embeddings.
-        Get item embeddings.
-        Pass user sessions through transformer blocks.
-
-        Parameters
-        ----------
-        sessions : torch.Tensor
-            User sessions in the form of sequences of items ids.
-
-        Returns
-        -------
-        (torch.Tensor, torch.Tensor)
-        """
-        item_embs = self.item_model.get_all_embeddings()  # [n_items + n_item_extra_tokens, n_factors]
-        session_embs = self.encode_sessions(sessions, item_embs)  # [batch_size, session_max_len, n_factors]
-        return item_embs, session_embs
-
-
-# ####  --------------  Lightning Model  --------------  #### #
-
-
-class TransformerLightningModuleBase(LightningModule):  # pylint: disable=too-many-instance-attributes
-    """
-    Base class for transfofmers lightning module. To change train procedure inherit
-    from this class and pass your custom LightningModule to your model parameters.
-
-    Parameters
-    ----------
-    torch_model : TransformerTorchBackbone
-        Torch model to make recommendations.
-    lr : float
-        Learning rate.
-    loss : str, default "softmax"
-        Loss function.
-    adam_betas : Tuple[float, float], default (0.9, 0.98)
-        Coefficients for running averages of gradient and its square.
-    data_preparator : TransformerDataPreparatorBase
-        Data preparator.
-    verbose : int, default 0
-        Verbosity level.
-    train_loss_name : str, default "train_loss"
-        Name of the training loss.
-    val_loss_name : str, default "val_loss"
-        Name of the training loss.
-    """
-
-    def __init__(
-        self,
-        torch_model: TransformerTorchBackbone,
-        model_config: tp.Dict[str, tp.Any],
-        dataset_schema: DatasetSchemaDict,
-        item_external_ids: ExternalIds,
-        data_preparator: TransformerDataPreparatorBase,
-        lr: float,
-        gbce_t: float,
-        loss: str,
-        verbose: int = 0,
-        train_loss_name: str = "train_loss",
-        val_loss_name: str = "val_loss",
-        adam_betas: tp.Tuple[float, float] = (0.9, 0.98),
-    ):
-        super().__init__()
-        self.torch_model = torch_model
-        self.model_config = model_config
-        self.dataset_schema = dataset_schema
-        self.item_external_ids = item_external_ids
-        self.lr = lr
-        self.loss = loss
-        self.adam_betas = adam_betas
-        self.gbce_t = gbce_t
-        self.data_preparator = data_preparator
-        self.verbose = verbose
-        self.train_loss_name = train_loss_name
-        self.val_loss_name = val_loss_name
-        self.item_embs: torch.Tensor
-
-        self.save_hyperparameters(ignore=["torch_model", "data_preparator"])
-
-    def configure_optimizers(self) -> torch.optim.Adam:
-        """Choose what optimizers and learning-rate schedulers to use in optimization"""
-        optimizer = torch.optim.Adam(self.torch_model.parameters(), lr=self.lr, betas=self.adam_betas)
-        return optimizer
-
-    def training_step(self, batch: tp.Dict[str, torch.Tensor], batch_idx: int) -> torch.Tensor:
-        """Training step."""
-        raise NotImplementedError()
-
-    def validation_step(self, batch: tp.Dict[str, torch.Tensor], batch_idx: int) -> tp.Dict[str, torch.Tensor]:
-        """Validate step."""
-        raise NotImplementedError()
-
-    def recommend_u2i(
-        self,
-        user_ids: InternalIdsArray,
-        recommend_dataloader: DataLoader,
-        sorted_item_ids_to_recommend: InternalIdsArray,
-        k: int,
-        ui_csr_for_filter: tp.Optional[sparse.csr_matrix],
-        *args: tp.Any,
-        **kwargs: tp.Any,
-    ) -> InternalRecoTriplet:
-        """Recommending."""
-        raise NotImplementedError()
-
-    def recommend_i2i(
-        self,
-        target_ids: InternalIdsArray,
-        sorted_item_ids_to_recommend: InternalIdsArray,
-        k: int,
-        recommend_n_threads: int,
-        recommend_use_gpu_ranking: bool,
-        recommend_device: tp.Optional[str],
-        *args: tp.Any,
-        **kwargs: tp.Any,
-    ) -> InternalRecoTriplet:
-        """Recommending to items."""
-        raise NotImplementedError()
-
-
-class TransformerLightningModule(TransformerLightningModuleBase):
-    """Lightning module to train transformer models."""
-
-    i2i_dist = Distance.COSINE
-
-    def on_train_start(self) -> None:
-        """Initialize parameters with values from Xavier normal distribution."""
-        self._xavier_normal_init()
-
-    def training_step(self, batch: tp.Dict[str, torch.Tensor], batch_idx: int) -> torch.Tensor:
-        """Training step."""
-        x, y, w = batch["x"], batch["y"], batch["yw"]
-        if self.loss == "softmax":
-            logits = self._get_full_catalog_logits(x)
-            loss = self._calc_softmax_loss(logits, y, w)
-        elif self.loss == "BCE":
-            negatives = batch["negatives"]
-            logits = self._get_pos_neg_logits(x, y, negatives)
-            loss = self._calc_bce_loss(logits, y, w)
-        elif self.loss == "gBCE":
-            negatives = batch["negatives"]
-            logits = self._get_pos_neg_logits(x, y, negatives)
-            loss = self._calc_gbce_loss(logits, y, w, negatives)
-        else:
-            loss = self._calc_custom_loss(batch, batch_idx)
-
-        self.log(self.train_loss_name, loss, on_step=False, on_epoch=True, prog_bar=self.verbose > 0)
-
-        return loss
-
-    def _calc_custom_loss(self, batch: tp.Dict[str, torch.Tensor], batch_idx: int) -> torch.Tensor:
-        raise ValueError(f"loss {self.loss} is not supported")
-
-    def on_validation_start(self) -> None:
-        """Save item embeddings"""
-        self.eval()
-        with torch.no_grad():
-            self.item_embs = self.torch_model.item_model.get_all_embeddings()
-
-    def on_validation_end(self) -> None:
-        """Clear item embeddings"""
-        del self.item_embs
-        torch.cuda.empty_cache()
-
-    def validation_step(self, batch: tp.Dict[str, torch.Tensor], batch_idx: int) -> tp.Dict[str, torch.Tensor]:
-        """Validate step."""
-        # x: [batch_size, session_max_len]
-        # y: [batch_size, 1]
-        # yw: [batch_size, 1]
-        x, y, w = batch["x"], batch["y"], batch["yw"]
-        outputs = {}
-        if self.loss == "softmax":
-            logits = self._get_full_catalog_logits(x)[:, -1:, :]
-            outputs["loss"] = self._calc_softmax_loss(logits, y, w)
-            outputs["logits"] = logits.squeeze()
-        elif self.loss == "BCE":
-            negatives = batch["negatives"]
-            pos_neg_logits = self._get_pos_neg_logits(x, y, negatives)[:, -1:, :]
-            outputs["loss"] = self._calc_bce_loss(pos_neg_logits, y, w)
-            outputs["pos_neg_logits"] = pos_neg_logits.squeeze()
-        elif self.loss == "gBCE":
-            negatives = batch["negatives"]
-            pos_neg_logits = self._get_pos_neg_logits(x, y, negatives)[:, -1:, :]
-            outputs["loss"] = self._calc_gbce_loss(pos_neg_logits, y, w, negatives)
-            outputs["pos_neg_logits"] = pos_neg_logits.squeeze()
-        else:
-            raise ValueError(f"loss {self.loss} is not supported")
-
-        self.log(self.val_loss_name, outputs["loss"], on_step=False, on_epoch=True, prog_bar=self.verbose > 0)
-        return outputs
-
-    def _get_full_catalog_logits(self, x: torch.Tensor) -> torch.Tensor:
-        item_embs, session_embs = self.torch_model(x)
-        logits = session_embs @ item_embs.T
-        return logits
-
-    def _get_pos_neg_logits(self, x: torch.Tensor, y: torch.Tensor, negatives: torch.Tensor) -> torch.Tensor:
-        # [n_items + n_item_extra_tokens, n_factors], [batch_size, session_max_len, n_factors]
-        item_embs, session_embs = self.torch_model(x)
-        pos_neg = torch.cat([y.unsqueeze(-1), negatives], dim=-1)  # [batch_size, session_max_len, n_negatives + 1]
-        pos_neg_embs = item_embs[pos_neg]  # [batch_size, session_max_len, n_negatives + 1, n_factors]
-        # [batch_size, session_max_len, n_negatives + 1]
-        logits = (pos_neg_embs @ session_embs.unsqueeze(-1)).squeeze(-1)
-        return logits
-
-    def _get_reduced_overconfidence_logits(self, logits: torch.Tensor, n_items: int, n_negatives: int) -> torch.Tensor:
-        # https://arxiv.org/pdf/2308.07192.pdf
-        alpha = n_negatives / (n_items - 1)  # sampling rate
-        beta = alpha * (self.gbce_t * (1 - 1 / alpha) + 1 / alpha)
-
-        pos_logits = logits[:, :, 0:1].to(torch.float64)
-        neg_logits = logits[:, :, 1:].to(torch.float64)
-
-        epsilon = 1e-10
-        pos_probs = torch.clamp(torch.sigmoid(pos_logits), epsilon, 1 - epsilon)
-        pos_probs_adjusted = torch.clamp(pos_probs.pow(-beta), 1 + epsilon, torch.finfo(torch.float64).max)
-        pos_probs_adjusted = torch.clamp(
-            torch.div(1, (pos_probs_adjusted - 1)), epsilon, torch.finfo(torch.float64).max
-        )
-        pos_logits_transformed = torch.log(pos_probs_adjusted)
-        logits = torch.cat([pos_logits_transformed, neg_logits], dim=-1)
-        return logits
-
-    @classmethod
-    def _calc_softmax_loss(cls, logits: torch.Tensor, y: torch.Tensor, w: torch.Tensor) -> torch.Tensor:
-        # We are using CrossEntropyLoss with a multi-dimensional case
-
-        # Logits must be passed in form of [batch_size, n_items + n_item_extra_tokens, session_max_len],
-        #  where n_items + n_item_extra_tokens is number of classes
-
-        # Target label indexes must be passed in a form of [batch_size, session_max_len]
-        # (`0` index for "PAD" ix excluded from loss)
-
-        # Loss output will have a shape of [batch_size, session_max_len]
-        # and will have zeros for every `0` target label
-        loss = torch.nn.functional.cross_entropy(
-            logits.transpose(1, 2), y, ignore_index=0, reduction="none"
-        )  # [batch_size, session_max_len]
-        loss = loss * w
-        n = (loss > 0).to(loss.dtype)
-        loss = torch.sum(loss) / torch.sum(n)
-        return loss
-
-    @classmethod
-    def _calc_bce_loss(cls, logits: torch.Tensor, y: torch.Tensor, w: torch.Tensor) -> torch.Tensor:
-        mask = y != 0
-        target = torch.zeros_like(logits)
-        target[:, :, 0] = 1
-
-        loss = torch.nn.functional.binary_cross_entropy_with_logits(
-            logits, target, reduction="none"
-        )  # [batch_size, session_max_len, n_negatives + 1]
-        loss = loss.mean(-1) * mask * w  # [batch_size, session_max_len]
-        loss = torch.sum(loss) / torch.sum(mask)
-        return loss
-
-    def _calc_gbce_loss(
-        self, logits: torch.Tensor, y: torch.Tensor, w: torch.Tensor, negatives: torch.Tensor
-    ) -> torch.Tensor:
-        n_actual_items = self.torch_model.item_model.n_items - self.data_preparator.n_item_extra_tokens
-        n_negatives = negatives.shape[2]
-        logits = self._get_reduced_overconfidence_logits(logits, n_actual_items, n_negatives)
-        loss = self._calc_bce_loss(logits, y, w)
-        return loss
-
-    def _xavier_normal_init(self) -> None:
-        for _, param in self.torch_model.named_parameters():
-            if param.data.dim() > 1:
-                torch.nn.init.xavier_normal_(param.data)
-
-    def _prepare_for_inference(self, recommend_device: tp.Optional[str]) -> None:
-        if recommend_device is None:
-            recommend_device = "cuda" if torch.cuda.is_available() else "cpu"
-        device = torch.device(recommend_device)
-        self.torch_model.to(device)
-        self.torch_model.eval()
-
-    def get_user_item_embeddings(
-        self,
-        recommend_dataloader: DataLoader,
-        recommend_device: tp.Optional[str],
-    ) -> tp.Tuple[torch.Tensor, torch.Tensor]:
-        """TODO."""
-        self._prepare_for_inference(recommend_device)
-        device = self.torch_model.item_model.device
-
-        with torch.no_grad():
-            item_embs = self.torch_model.item_model.get_all_embeddings()
-            user_embs = []
-            for batch in recommend_dataloader:
-                batch_embs = self.torch_model.encode_sessions(batch["x"].to(device), item_embs)[:, -1, :]
-                user_embs.append(batch_embs)
-
-        return torch.cat(user_embs), item_embs
-
-    def recommend_u2i(
-        self,
-        user_ids: InternalIdsArray,
-        recommend_dataloader: DataLoader,
-        sorted_item_ids_to_recommend: InternalIdsArray,
-        k: int,
-        ui_csr_for_filter: tp.Optional[sparse.csr_matrix],
-        recommend_n_threads: int,
-        recommend_use_gpu_ranking: bool,
-        recommend_device: tp.Optional[str],
-    ) -> InternalRecoTriplet:
-        """TODO."""
-        user_embs, item_embs = self.get_user_item_embeddings(recommend_dataloader, recommend_device)
-        user_embs_np = user_embs.detach().cpu().numpy()
-        item_embs_np = item_embs.detach().cpu().numpy()
-        del user_embs, item_embs
-        torch.cuda.empty_cache()
-
-        ranker = ImplicitRanker(
-            Distance.DOT,
-            user_embs_np[user_ids],  # [n_rec_users, n_factors]
-            item_embs_np,  # [n_items + n_item_extra_tokens, n_factors]
-        )
-
-        if ui_csr_for_filter is not None:
-            ui_csr_for_filter = ui_csr_for_filter[user_ids]
-
-        # TODO: We should test if torch `topk`` is faster when `filter_viewed`` is ``False``
-        user_ids_indices, all_reco_ids, all_scores = ranker.rank(
-            subject_ids=np.arange(user_embs_np.shape[0]),  # n_rec_users
-            k=k,
-            filter_pairs_csr=ui_csr_for_filter,  # [n_rec_users x n_items + n_item_extra_tokens]
-            sorted_object_whitelist=sorted_item_ids_to_recommend,  # model_internal
-            num_threads=recommend_n_threads,
-            use_gpu=recommend_use_gpu_ranking and HAS_CUDA,
-        )
-        all_user_ids = user_ids[user_ids_indices]
-        return all_user_ids, all_reco_ids, all_scores
-
-    def recommend_i2i(
-        self,
-        target_ids: InternalIdsArray,
-        sorted_item_ids_to_recommend: InternalIdsArray,
-        k: int,
-        recommend_n_threads: int,
-        recommend_use_gpu_ranking: bool,
-        recommend_device: tp.Optional[str],
-    ) -> InternalRecoTriplet:
-        """TODO."""
-        self._prepare_for_inference(recommend_device)
-
-        with torch.no_grad():
-            item_embs = self.torch_model.item_model.get_all_embeddings().detach().cpu().numpy()
-        # TODO: i2i recommendations do not need filtering viewed and user most of the times has GPU
-        # We should test if torch `topk`` is faster
-
-        ranker = ImplicitRanker(
-            self.i2i_dist,
-            item_embs,  # [n_items + n_item_extra_tokens, n_factors]
-            item_embs,  # [n_items + n_item_extra_tokens, n_factors]
-        )
-        return ranker.rank(
-            subject_ids=target_ids,  # model internal
-            k=k,
-            filter_pairs_csr=None,
-            sorted_object_whitelist=sorted_item_ids_to_recommend,  # model internal
-            num_threads=recommend_n_threads,
-            use_gpu=recommend_use_gpu_ranking and HAS_CUDA,
-        )
-
 
 # ####  --------------  Transformer Model Config  --------------  #### #
 
@@ -934,18 +376,14 @@ class TransformerModelBase(ModelBase[TransformerModelConfig_T]):  # pylint: disa
         if sorted_item_ids_to_recommend is None:
             sorted_item_ids_to_recommend = self.data_preparator.get_known_items_sorted_internal_ids()  # model internal
 
-        ui_csr_for_filter = None
-        if filter_viewed:
-            ui_csr_for_filter = dataset.get_user_item_matrix(include_weights=False)
-
         recommend_dataloader = self.data_preparator.get_dataloader_recommend(dataset, self.recommend_batch_size)
-
         return self.lightning_model.recommend_u2i(
             user_ids=user_ids,
             recommend_dataloader=recommend_dataloader,
             sorted_item_ids_to_recommend=sorted_item_ids_to_recommend,
             k=k,
-            ui_csr_for_filter=ui_csr_for_filter,
+            filter_viewed=filter_viewed,
+            dataset=dataset,
             recommend_n_threads=self.recommend_n_threads,
             recommend_use_gpu_ranking=self.recommend_use_gpu_ranking,
             recommend_device=self.recommend_device,
