+++ tests/models/nn/test_bert4rec.py
@@ -730,7 +730,7 @@ class TestBERT4RecModelConfiguration:
                 "pos_encoding_type": "rectools.models.nn.transformer_net_blocks.LearnableInversePositionalEncoding",
                 "transformer_layers_type": "rectools.models.nn.transformer_net_blocks.PreLNTransformerLayers",
                 "data_preparator_type": "rectools.models.nn.bert4rec.BERT4RecDataPreparator",
-                "lightning_module_type": "rectools.models.nn.transformer_base.TransformerLightningModule",
+                "lightning_module_type": "rectools.models.nn.transformer_lightning.TransformerLightningModule",
                 "get_val_mask_func": "tests.models.nn.utils.leave_one_out_mask",
             }
             expected.update(simple_types_params)
